Camera game objects display render to window

+++ AATelloEngine/W_DockWindow.cpp
@@ -111,7 +111,7 @@ void W_DockWindow::DrawFileMenu()
 			App->scene->SaveScene();
 
 		if (ImGui::MenuItem("Load"))
-			App->editor->OpenWindow(E_WINDOW_TYPE::LOAD_FILE);
+			App->editor->OpenWindow((int)E_WINDOW_TYPE::LOAD_FILE);
 
 		if (ImGui::MenuItem("Load Scene"))
 			App->scene->LoadScene();
